requirements.txt
flask
spacy
PyPDF2
python-docx

